requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
ollama>=0.1.0
pydantic>=2.0.0
PyPDF2>=3.0.0
pypdf>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
reportlab>=4.0.0
pymupdf>=1.23.0
numpy>=1.24.0,<2.0
python-dotenv>=1.0.0
# Dependențe pentru sistemul OCR
paddleocr>=2.7.0
paddlepaddle>=3.0.0
opencv-python>=4.8.0
python-docx>=1.1.0
mysql-connector-python>=8.2.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-multipart>=0.0.6

